fix: update isFunded logic based on token balances
